requirements.txt
# App packages
requests==2.6.0
fcache==0.4.7
mmh3==2.5.1
APScheduler==3.6.1
python-dateutil==2.4.2
chainmap==1.0.3

# Development packages
bumpversion
coveralls
mimesis==2.0.1
mkdocs
mypy
pur
pylint
pytest
pytest-cov
pytest-flake8
pytest-html
pytest-rerunfailures
pytest-runner
pytest-xdist
responses
tox
twine
